requirements.txt
flask
flask-smorest
python-dotenv
flask-marshmallow
Flask-SQLAlchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
requests
psycopg2-binary
rq
Jinja2